refactor : Modal 폴더안 Modal컴포넌트 파일명을 index.ts로 변경
