+++ components/src/stories/modal.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Modal from '../Modal/Modal';
+import Modal from '../Modal';
 
 const meta: Meta<typeof Modal> = {
   title: 'Components/Modal',
